Support HiDPI on Linux by querying the Xft DPI through XSETTINGS.
The Xft DPI is set when the user adjusts the scale in the Ubuntu
"Displays" preference pane.

This depends on `rust-xsettings`, which in turn uses the C library
`libxsettings-client-0`. Unfortunately, this library is not installed by
default on most Linux distributions. To avoid unnecessary dependencies,
this commit places support for HiDPI on Linux behind the `use-xsettings`
Cargo feature.